requirements.txt
# --- kompatible Trio für CUDA ≥12.1 ---
torch==2.4.0
torchvision==0.19.0
torchaudio==2.4.0

# --- aktuelle Libraries ---
accelerate==0.34.0
git+https://github.com/huggingface/diffusers
transformers==4.45 # 4.43.1
deepspeed==0.14.2
einops==0.7.0
timm==0.9.16
insightface==0.7.3

# --- Utils ---
opencv-python-headless==4.10.0.82
pillow==10.3.0
scikit-image==0.23.2
scikit-learn==1.5.0
tqdm==4.66.4
sentencepiece==0.2.0
hf_transfer
